requirements.txt
Flask
Flask-PyMongo
pymongo
python-dotenv
pymongo[srv]